fix(firestore, web): check blob constructor
