+++ packages/cloud_firestore/cloud_firestore_web/lib/src/document_reference_web.dart
@@ -6,7 +6,7 @@ import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_inte
 
 import 'internals.dart';
 import 'utils/web_utils.dart';
-import 'utils/codec_utility.dart';
+import 'utils/encode_utility.dart';
 import 'interop/firestore.dart' as firestore_interop;
 
 /// Web implementation for Firestore [DocumentReferencePlatform].
