+++ packages/cloud_firestore/cloud_firestore_web/lib/src/field_value_factory_web.dart
@@ -6,7 +6,7 @@ import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_inte
 import 'package:js/js_util.dart';
 
 import 'field_value_web.dart';
-import 'utils/codec_utility.dart';
+import 'utils/encode_utility.dart';
 import 'interop/firestore.dart' as firestore_interop;
 
 /// An implementation of [FieldValueFactoryPlatform] which builds [FieldValuePlatform]
