+++ packages/cloud_firestore/cloud_firestore_web/lib/src/query_web.dart
@@ -3,7 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-import 'package:cloud_firestore_web/src/utils/codec_utility.dart';
+import 'package:cloud_firestore_web/src/utils/encode_utility.dart';
 
 import 'internals.dart';
 import 'interop/firestore.dart' as firestore_interop;
