+++ packages/cloud_firestore/cloud_firestore_web/lib/src/transaction_web.dart
@@ -6,7 +6,7 @@ import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_inte
 
 import 'internals.dart';
 import 'interop/firestore.dart' as firestore_interop;
-import 'utils/codec_utility.dart';
+import 'utils/encode_utility.dart';
 import 'utils/web_utils.dart';
 
 /// A web specific implementation of [Transaction].
