+++ packages/cloud_firestore/cloud_firestore_web/lib/src/utils/encode_utility.dart
@@ -3,8 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-import 'package:cloud_firestore_web/cloud_firestore_web.dart'
-    show FirebaseFirestoreWeb;
 
 import '../interop/firestore.dart' as firestore_interop;
 import '../document_reference_web.dart';
@@ -110,52 +108,15 @@ class CodecUtility {
     } else if (value is Timestamp) {
       return value.toDate();
     } else if (value is GeoPoint) {
-      return firestore_interop.GeoPoint(value.latitude, value.longitude);
+      return firestore_interop.GeoPointJsImpl(value.latitude, value.longitude);
     } else if (value is Blob) {
-      return firestore_interop.Blob.fromUint8Array(value.bytes);
+      return firestore_interop.BlobJsImpl.fromUint8Array(value.bytes);
     } else if (value is DocumentReferenceWeb) {
       return value.firestoreWeb.doc(value.path);
     } else if (value is Map<String, dynamic>) {
       return encodeMapData(value);
     } else if (value is List<dynamic>) {
       return encodeArrayData(value);
-    }
-    return value;
-  }
-
-  /// Decodes the values on an incoming Map to their proper types.
-  static Map<String, dynamic>? decodeMapData(Map<String, dynamic>? data) {
-    if (data == null) {
-      return null;
-    }
-    Map<String, dynamic> output = Map.from(data);
-    output.updateAll((key, value) => valueDecode(value));
-    return output;
-  }
-
-  /// Decodes the values on an incoming Array to their proper types.
-  static List<dynamic>? decodeArrayData(List<dynamic>? data) {
-    if (data == null) {
-      return null;
-    }
-    return List.from(data).map(valueDecode).toList();
-  }
-
-  /// Decodes an incoming value to its proper type.
-  static dynamic valueDecode(dynamic value) {
-    if (value.latitude != null && value.longitude != null) {
-      return GeoPoint(value.latitude as double, value.longitude as double);
-    } else if (value is DateTime) {
-      return Timestamp.fromDate(value);
-    } else if (value is firestore_interop.Blob) {
-      return Blob(value.toUint8Array());
-    } else if (value is firestore_interop.DocumentReference) {
-      return (FirebaseFirestorePlatform.instance as FirebaseFirestoreWeb)
-          .doc(value.path);
-    } else if (value is Map<String, dynamic>) {
-      return decodeMapData(value);
-    } else if (value is List<dynamic>) {
-      return decodeArrayData(value);
     }
     return value;
   }
